Fixed PR remarks
