+++ apps/charge-smart-wallets-service/src/data-layer/data-layer.module.ts
@@ -3,12 +3,13 @@ import { DataLayerService } from '@app/smart-wallets-service/data-layer/data-lay
 import { dataLayerProviders } from '@app/smart-wallets-service/data-layer/data-layer.providers'
 import { Module } from '@nestjs/common'
 import { DatabaseModule } from '@app/common'
-import { UserOpFactory } from '@app/smart-wallets-service/common/utils/user-op-parser'
+import { UserOpFactory } from '@app/smart-wallets-service/common/services/user-op-factory.service'
+import { UserOpParser } from '@app/smart-wallets-service/common/services/user-op-parser.service'
 
 @Module({
   imports: [DatabaseModule],
   controllers: [DataLayerController],
-  providers: [DataLayerService, ...dataLayerProviders, UserOpFactory]
+  providers: [DataLayerService, ...dataLayerProviders, UserOpFactory, UserOpParser]
 })
 
 export class DataLayerModule { }
